refactor: Separate evaluator concerns into focused modules
Split evaluator/core.rs (769 lines) into layered modules:
- caching.rs: CSE and subquery cache configuration
- context.rs: Trigger/procedural/CTE context management
- parallel.rs: Parallel execution support
- single.rs: ExpressionEvaluator implementation
- combined_core.rs: CombinedExpressionEvaluator implementation
- core.rs: Shared evaluation logic only

Benefits:
- Clearer responsibilities: evaluation vs caching vs context
- Easier caching improvements: isolated cache strategies
- Better context handling: separate context resolution
- Reduced complexity: each module handles one concern

All existing tests pass (1127/1132, 5 pre-existing failures confirmed).
No performance regression - caching still works via environment variables.

Closes #2387
