+++ crates/vibesql-executor/src/evaluator/mod.rs
@@ -3,6 +3,11 @@ pub(crate) mod casting;
 pub mod coercion;
 mod combined;
 mod core;
+mod caching;
+mod context;
+mod parallel;
+mod single;
+mod combined_core;
 pub mod date_format;
 mod expression_hash;
 mod expressions;
